Separate file system watching from IFileService file
