+++ src/Desktop/Services/FileSystemChangeType.cs
@@ -1,0 +1,9 @@
+namespace Desktop.Services;
+
+public enum FileSystemChangeType
+{
+    Created,
+    Deleted,
+    Changed,
+    Renamed
+}
